hide some implementation details
